update - inserindo atualização de kpi
