+++ site/public/dashboard-espec.html
@@ -20,7 +20,7 @@
 
     <title>Dashboard Específico</title>
   </head>
-  <body>
+  <body onload="iniciar()">
     <div class="container-menu">
 
       <div class="img-logo">
@@ -144,7 +144,7 @@
                     <li>
                       <span class="circle-kpi blue"></span>
 
-                      <span class="text-kpi">Perigo (< 18%)</span>
+                      <span class="text-kpi">Perigo ( 18%)</span>
                     </li>
 
                     <li>
@@ -270,7 +270,7 @@
                     <li>
                       <span class="circle-kpi blue"></span>
 
-                      <span class="text-kpi">Perigo (< 18%)</span>
+                      <span class="text-kpi">Perigo ( 18%)</span>
                     </li>
 
                     <li>
@@ -423,7 +423,7 @@
 </html>
 
 <script src="./js/dashboard-espec.js"></script>
-
+<script src="./js/funcoes-maquinas.js"></script>
 <script
   type="module"
   src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"
@@ -432,3 +432,7 @@
   nomodule
   src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"
 ></script>
+
+<script>
+  
+</script>
